fix: do not filer on end date of mandataris for notification of bekrachtigde
We filtered on the end date of the mandataris to be before or equal to today.
As the end date in this period is often set to 2030 these mandatarissen will be filtered out.

Removed this filter check as we actually do not care about the end date. When a mandataris is started it needs to be bekrachtigd.

Refs: lmb-1744